chore: retire @vizij/arora-types (fold values into @vizij/value-json)
The @vizij/arora-types package was a hand-maintained TS mirror in two halves,
both now redundant:

- Value serde half (AroraValue/AroraType, extractNumericValue, the f64/… value
  constructors and guards): superseded by @vizij/value-json, which now carries
  the Rust-generated arora Value bindings (AroraValueJSON) and the valueAs*
  accessors that decode them.
- WebSocket message-protocol half (SlotInfo/NodeInfo/Incoming/Outgoing and the
  create*/List*/Invoke constructors): the shared in-repo arora-connection /
  websocket protocol it mirrored has been retired, so these are app-specific now.

Its only consumer was apps/vizij-standalone useWebSocketSync.ts, so:

- extractNumericValue -> valueAsNumber (already imported from value-json; a
  superset that decodes every arora numeric tag and returns undefined for
  non-numeric — the null check became an undefined check).
- AroraValue -> AroraValueJSON from @vizij/value-json.
- f64(x) constructor -> inline arora-serde literal { f64: x }.
- NodeInfo + AroraType (the only protocol pieces the hook uses) moved into a
  new app-local apps/vizij-standalone/src/hooks/aroraWsProtocol.ts, trimmed to
  the slot-descriptor shapes the native set_slots command actually needs.

Package directory, the workspace dependency, and the README / AGENTS workspace
listings are removed; pnpm-lock.yaml regenerated.

Verified: apps/vizij-standalone vite build passes (1056 modules, clean); tsc on
the migrated files (useWebSocketSync.ts, aroraWsProtocol.ts) reports no errors
(the 3 pre-existing App.tsx import.meta.env / unused-local errors are unrelated
and present on origin/main); prettier --check clean on the edited docs.
