+++ apps/vizij-standalone/package.json
@@ -17,7 +17,6 @@
     "@tauri-apps/plugin-dialog": "2.6.0",
     "@tauri-apps/plugin-fs": "2.4.5",
     "@tauri-apps/plugin-log": "2.8.0",
-    "@vizij/arora-types": "workspace:*",
     "@vizij/node-graph-authoring": "workspace:*",
     "@vizij/node-graph-react": "workspace:*",
     "@vizij/node-graph-wasm": "^0.7.0",
