--- packages/@vizij/arora-types/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@vizij/arora-types",
-  "description": "TypeScript types for arora-types values and types. Matches Rust arora-types serde format.",
-  "version": "0.1.0",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "sideEffects": false,
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    }
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/vizij-ai/vizij-web",
-    "directory": "packages/@vizij/arora-types"
-  },
-  "license": "MIT",
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsup src/index.ts --format esm,cjs --dts src/index.ts",
-    "dev": "tsup src/index.ts --format esm,cjs --watch --dts src/index.ts",
-    "test": "vitest --run",
-    "typecheck": "tsc --noEmit",
-    "lint": "pnpm --filter \"$npm_package_name\" exec eslint --ext .js,.jsx,.ts,.tsx -- .",
-    "lint:fix": "pnpm --filter \"$npm_package_name\" exec eslint --ext .js,.jsx,.ts,.tsx --fix -- .",
-    "clean": "rm -rf dist .turbo coverage tsconfig.tsbuildinfo",
-    "reset": "rm -rf node_modules",
-    "prepack": "pnpm run build && pnpm run typecheck"
-  },
-  "devDependencies": {
-    "tsup": "^8.0.1",
-    "typescript": "^5.5.0",
-    "vitest": "^3.2.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {}
-}
