--- packages/@vizij/arora-types/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ES2020",
-    "declaration": true,
-    "outDir": "dist",
-    "strict": true,
-    "moduleResolution": "Bundler",
-    "skipLibCheck": true
-  },
-  "include": ["src"],
-  "exclude": ["dist", "build", "node_modules"]
-}
